requirements.txt
# Core FastAPI and server
fastapi==0.111.0           # Updated to support anyio >= 4.8.0
uvicorn==0.29.0            # Latest stable, compatible
pydantic==2.5.0
pydantic[email]==2.5.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
# Database
motor==3.3.2
pymongo==4.6.0

# Environment and config
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0
httpx
# AI/ML APIs
openai
google-genai==1.24.0       # Requires anyio >= 4.8.0

# Logging and monitoring
structlog==23.2.0

# Email utilities
emails==0.6.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
